run-test-plan: fail when test plan creation fails etc
A recent typo meant some test plans weren't running, and the error was
just been logged and ignored, and because the plan hadn't even been
created it never made it into the 'analysis' part of the script to be
flagged as an issue.

Just exit when we get an error. This resolves the issue as seen here:

https://gitlab.com/openid/conformance-suite/-/jobs/2293588330

(I suspect I'm not quite handling this right, and should perhaps be
draining the exception from the queue somehow. But this seems to work.)